+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableByteArray.kt
@@ -553,7 +553,7 @@ public value class ImmutableByteArray @PublishedApi internal constructor(
 
         var minValue = selector(minElement)
         for (i in 1..lastIndex) {
-            val currentElement = this[i]
+            val currentElement = values[i]
             val currentValue = selector(currentElement)
             if (currentValue < minValue) {
                 minElement = currentElement
@@ -561,6 +561,26 @@ public value class ImmutableByteArray @PublishedApi internal constructor(
             }
         }
         return minElement
+    }
+
+    /**
+     * @return the first element which the [selector] yields the largest value.
+     * @throws NoSuchElementException if this ImmutableByteArray is empty
+     */
+    public inline fun <R : Comparable<R>> maxBy(selector: (element: Byte) -> R): Byte {
+        var maxElement = first()
+        if (size == 1) return maxElement
+
+        var maxValue = selector(maxElement)
+        for (i in 1..lastIndex) {
+            val currentElement = values[i]
+            val currentValue = selector(currentElement)
+            if (currentValue > maxValue) {
+                maxElement = currentElement
+                maxValue = currentValue
+            }
+        }
+        return maxElement
     }
 
     /**